+++ src/tables/nonconst_column.cpp
@@ -53,7 +53,7 @@ namespace tables
         add_rows(1);
     }
 
-    size_t nonconst_column::rows() const
+    uint64_t nonconst_column::rows() const
     {
         return m_values.size();
     }
@@ -102,7 +102,7 @@ namespace tables
         m_default_value = value;
     }
 
-    boost::optional<size_t> nonconst_column::type_hash() const
+    boost::optional<uint64_t> nonconst_column::type_hash() const
     {
         return m_type_hash;
     }
